+++ linode/nbnode/framework_datasource.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
-	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/linode/linodego"
 	"github.com/linode/terraform-provider-linode/linode/helper"
 )
@@ -16,17 +15,6 @@ func NewDataSource() datasource.DataSource {
 
 type DataSource struct {
 	client *linodego.Client
-}
-
-func (data *DataSourceModel) parseNodeBalancerNode(nbnode *linodego.NodeBalancerNode) {
-	data.ID = types.Int64Value(int64(nbnode.ID))
-	data.NodeBalancerID = types.Int64Value(int64(nbnode.NodeBalancerID))
-	data.ConfigID = types.Int64Value(int64(nbnode.ConfigID))
-	data.Label = types.StringValue(nbnode.Label)
-	data.Weight = types.Int64Value(int64(nbnode.Weight))
-	data.Mode = types.StringValue(string(nbnode.Mode))
-	data.Address = types.StringValue(nbnode.Address)
-	data.Status = types.StringValue(nbnode.Status)
 }
 
 func (d *DataSource) Configure(
@@ -45,17 +33,6 @@ func (d *DataSource) Configure(
 	}
 
 	d.client = meta.Client
-}
-
-type DataSourceModel struct {
-	ID             types.Int64  `tfsdk:"id"`
-	NodeBalancerID types.Int64  `tfsdk:"nodebalancer_id"`
-	ConfigID       types.Int64  `tfsdk:"config_id"`
-	Label          types.String `tfsdk:"label"`
-	Weight         types.Int64  `tfsdk:"weight"`
-	Mode           types.String `tfsdk:"mode"`
-	Address        types.String `tfsdk:"address"`
-	Status         types.String `tfsdk:"status"`
 }
 
 func (d *DataSource) Metadata(
@@ -100,6 +77,6 @@ func (d *DataSource) Read(
 		return
 	}
 
-	data.parseNodeBalancerNode(node)
+	data.ParseNodeBalancerNode(node)
 	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
 }
